move user models to domain/user.rs
