+++ src/domain/mod.rs
@@ -1,10 +1,12 @@
+mod user;
+pub use user::{NewUser, User};
+
 use std::marker::PhantomData;
 
 use argon2::{
     password_hash::{self, SaltString},
     Algorithm, Argon2, Params, PasswordHasher, Version,
 };
-use time::OffsetDateTime;
 use unicode_segmentation::UnicodeSegmentation;
 use validator::validate_email;
 
@@ -127,29 +129,5 @@ where
     match f(value) {
         true => Ok(()),
         false => Err(ValidationError(msg.into())),
-    }
-}
-
-pub struct NewUser {
-    pub username: Username,
-    pub email: Email,
-    pub password: Password,
-}
-
-pub struct User {
-    pub id: Id<Self>,
-    pub username: Username,
-    pub email: Email,
-    pub created_at: OffsetDateTime,
-    pub updated_at: OffsetDateTime,
-}
-
-impl User {
-    pub fn id(&self) -> u64 {
-        self.id.value
-    }
-    pub fn set_id(mut self, id: u64) -> Self {
-        self.id = id.into();
-        self
     }
 }
